improvement: switch bundle to use dist folder
- this is closer to what apps will consume
- a second rollup configuration is used to process less->css and use a watch mode
- remove theme bundle functionality, themes are inlined now
