+++ packages/main/.eslintignore
@@ -6,5 +6,5 @@ lib
 test
 bundle.esm.js
 bundle.es5.js
-rollup.config.js
+rollup.config*.js
 wdio.conf.js
