+++ packages/main/bundle.esm.js
@@ -3,7 +3,6 @@ import "@ui5/webcomponents-base/src/sap/ui/webcomponents/base/browsersupport/Edg
 
 import "@ui5/webcomponents-base/src/sap/ui/webcomponents/base/shims/jquery-shim";
 import "@ui5/webcomponents-base/src/sap/ui/webcomponents/base/events/PolymerGestures";
-import "./src/ThemeProvider";
 
 import Gregorian from "@ui5/webcomponents-core/dist/sap/ui/core/date/Gregorian";
 import Buddhist from "@ui5/webcomponents-core/dist/sap/ui/core/date/Buddhist";
@@ -11,45 +10,45 @@ import Islamic from "@ui5/webcomponents-core/dist/sap/ui/core/date/Islamic";
 import Japanese from "@ui5/webcomponents-core/dist/sap/ui/core/date/Japanese";
 import Persian from "@ui5/webcomponents-core/dist/sap/ui/core/date/Persian";
 
-import Button from "./src/Button";
-import CheckBox from "./src/CheckBox";
-import Card from "./src/Card";
-import Calendar from "./src/Calendar";
-import CalendarHeader from "./src/CalendarHeader";
-import DatePicker from "./src/DatePicker";
-import Dialog from "./src/Dialog";
-import Icon from "./src/Icon";
-import Input from "./src/Input";
-import InputSuggestions from "./src/InputSuggestions";
-import Label from "./src/Label";
-import Link from "./src/Link";
-import DayPicker from "./src/DayPicker";
-import Popover from "./src/Popover";
-import Panel from "./src/Panel";
-import RadioButton from "./src/RadioButton";
-import Select from "./src/Select";
-import ShellBar from "./src/ShellBar";
-import ShellBarItem from "./src/ShellBarItem";
-import Switch from "./src/Switch";
-import TabContainer from "./src/TabContainer";
-import Tab from "./src/Tab";
-import TabSeparator from "./src/TabSeparator";
-import Table from "./src/Table";
-import TableColumn from "./src/TableColumn";
-import TableRow from "./src/TableRow";
-import TableCell from "./src/TableCell";
-import TextArea from "./src/TextArea";
-import Timeline from "./src/Timeline";
-import TimelineItem from "./src/TimelineItem";
-import Title from "./src/Title";
-import ToggleButton from "./src/ToggleButton";
-import Toolbar from "./src/Toolbar";
-import ToolbarSpacer from "./src/ToolbarSpacer";
+import Button from "./dist/Button";
+import CheckBox from "./dist/CheckBox";
+import Card from "./dist/Card";
+import Calendar from "./dist/Calendar";
+import CalendarHeader from "./dist/CalendarHeader";
+import DatePicker from "./dist/DatePicker";
+import Dialog from "./dist/Dialog";
+import Icon from "./dist/Icon";
+import Input from "./dist/Input";
+import InputSuggestions from "./dist/InputSuggestions";
+import Label from "./dist/Label";
+import Link from "./dist/Link";
+import DayPicker from "./dist/DayPicker";
+import Popover from "./dist/Popover";
+import Panel from "./dist/Panel";
+import RadioButton from "./dist/RadioButton";
+import Select from "./dist/Select";
+import ShellBar from "./dist/ShellBar";
+import ShellBarItem from "./dist/ShellBarItem";
+import Switch from "./dist/Switch";
+import TabContainer from "./dist/TabContainer";
+import Tab from "./dist/Tab";
+import TabSeparator from "./dist/TabSeparator";
+import Table from "./dist/Table";
+import TableColumn from "./dist/TableColumn";
+import TableRow from "./dist/TableRow";
+import TableCell from "./dist/TableCell";
+import TextArea from "./dist/TextArea";
+import Timeline from "./dist/Timeline";
+import TimelineItem from "./dist/TimelineItem";
+import Title from "./dist/Title";
+import ToggleButton from "./dist/ToggleButton";
+import Toolbar from "./dist/Toolbar";
+import ToolbarSpacer from "./dist/ToolbarSpacer";
 
-import List from "./src/List";
-import StandardListItem from "./src/StandardListItem";
-import CustomListItem from "./src/CustomListItem";
-import GroupHeaderListItem from "./src/GroupHeaderListItem";
+import List from "./dist/List";
+import StandardListItem from "./dist/StandardListItem";
+import CustomListItem from "./dist/CustomListItem";
+import GroupHeaderListItem from "./dist/GroupHeaderListItem";
 
 // used in test pages
 import RenderScheduler from "@ui5/webcomponents-base/src/sap/ui/webcomponents/base/RenderScheduler";
