+++ packages/main/entries-less.esm.js
@@ -1,0 +1,151 @@
+import "./src/themes/sap_belize/Button.less";
+import "./src/themes/sap_belize_hcb/Button.less";
+import "./src/themes/sap_fiori_3/Button.less";
+
+import "./src/themes/sap_belize/CheckBox.less";
+import "./src/themes/sap_belize_hcb/CheckBox.less";
+import "./src/themes/sap_fiori_3/CheckBox.less";
+
+import "./src/themes/sap_belize/Card.less";
+import "./src/themes/sap_belize_hcb/Card.less";
+import "./src/themes/sap_fiori_3/Card.less";
+
+import "./src/themes/sap_belize/Calendar.less";
+import "./src/themes/sap_belize_hcb/Calendar.less";
+import "./src/themes/sap_fiori_3/Calendar.less";
+
+import "./src/themes/sap_belize/CalendarHeader.less";
+import "./src/themes/sap_belize_hcb/CalendarHeader.less";
+import "./src/themes/sap_fiori_3/CalendarHeader.less";
+
+import "./src/themes/sap_belize/DatePicker.less";
+import "./src/themes/sap_belize_hcb/DatePicker.less";
+import "./src/themes/sap_fiori_3/DatePicker.less";
+
+import "./src/themes/sap_belize/Dialog.less";
+import "./src/themes/sap_belize_hcb/Dialog.less";
+import "./src/themes/sap_fiori_3/Dialog.less";
+
+import "./src/themes/sap_belize/Icon.less";
+import "./src/themes/sap_belize_hcb/Icon.less";
+import "./src/themes/sap_fiori_3/Icon.less";
+
+import "./src/themes/sap_belize/Input.less";
+import "./src/themes/sap_belize_hcb/Input.less";
+import "./src/themes/sap_fiori_3/Input.less";
+
+import "./src/themes/sap_belize/Label.less";
+import "./src/themes/sap_belize_hcb/Label.less";
+import "./src/themes/sap_fiori_3/Label.less";
+
+import "./src/themes/sap_belize/ShellBarInput.less";
+import "./src/themes/sap_belize_hcb/ShellBarInput.less";
+import "./src/themes/sap_fiori_3/ShellBarInput.less";
+
+import "./src/themes/sap_belize/Link.less";
+import "./src/themes/sap_belize_hcb/Link.less";
+import "./src/themes/sap_fiori_3/Link.less";
+
+import "./src/themes/sap_belize/DayPicker.less";
+import "./src/themes/sap_belize_hcb/DayPicker.less";
+import "./src/themes/sap_fiori_3/DayPicker.less";
+
+import "./src/themes/sap_belize/Popover.less";
+import "./src/themes/sap_belize_hcb/Popover.less";
+import "./src/themes/sap_fiori_3/Popover.less";
+
+import "./src/themes/sap_belize/Panel.less";
+import "./src/themes/sap_belize_hcb/Panel.less";
+import "./src/themes/sap_fiori_3/Panel.less";
+
+import "./src/themes/sap_belize/RadioButton.less";
+import "./src/themes/sap_belize_hcb/RadioButton.less";
+import "./src/themes/sap_fiori_3/RadioButton.less";
+
+import "./src/themes/sap_belize/Switch.less";
+import "./src/themes/sap_belize_hcb/Switch.less";
+import "./src/themes/sap_fiori_3/Switch.less";
+
+import "./src/themes/sap_belize/Select.less";
+import "./src/themes/sap_belize_hcb/Select.less";
+import "./src/themes/sap_fiori_3/Select.less";
+
+import "./src/themes/sap_belize/ShellBar.less";
+import "./src/themes/sap_belize_hcb/ShellBar.less";
+import "./src/themes/sap_fiori_3/ShellBar.less";
+
+import "./src/themes/sap_belize/TabContainer.less";
+import "./src/themes/sap_belize_hcb/TabContainer.less";
+import "./src/themes/sap_fiori_3/TabContainer.less";
+
+import "./src/themes/sap_belize/Table.less";
+import "./src/themes/sap_belize_hcb/Table.less";
+import "./src/themes/sap_fiori_3/Table.less";
+
+import "./src/themes/sap_belize/TableRow.less";
+import "./src/themes/sap_belize_hcb/TableRow.less";
+import "./src/themes/sap_fiori_3/TableRow.less";
+
+import "./src/themes/sap_belize/TextArea.less";
+import "./src/themes/sap_belize_hcb/TextArea.less";
+import "./src/themes/sap_fiori_3/TextArea.less";
+
+import "./src/themes/sap_belize/TableCell.less";
+import "./src/themes/sap_belize_hcb/TableCell.less";
+import "./src/themes/sap_fiori_3/TableCell.less";
+
+import "./src/themes/sap_belize/TableColumn.less";
+import "./src/themes/sap_belize_hcb/TableColumn.less";
+import "./src/themes/sap_fiori_3/TableColumn.less";
+
+import "./src/themes/sap_belize/Title.less";
+import "./src/themes/sap_belize_hcb/Title.less";
+import "./src/themes/sap_fiori_3/Title.less";
+
+import "./src/themes/sap_belize/TimelineItem.less";
+import "./src/themes/sap_belize_hcb/TimelineItem.less";
+import "./src/themes/sap_fiori_3/TimelineItem.less";
+
+import "./src/themes/sap_belize/Timeline.less";
+import "./src/themes/sap_belize_hcb/Timeline.less";
+import "./src/themes/sap_fiori_3/Timeline.less";
+
+import "./src/themes/sap_belize/ToggleButton.less";
+import "./src/themes/sap_belize_hcb/ToggleButton.less";
+import "./src/themes/sap_fiori_3/ToggleButton.less";
+
+import "./src/themes/sap_belize/Toolbar.less";
+import "./src/themes/sap_belize_hcb/Toolbar.less";
+import "./src/themes/sap_fiori_3/Toolbar.less";
+
+import "./src/themes/sap_belize/ToolbarSpacer.less";
+import "./src/themes/sap_belize_hcb/ToolbarSpacer.less";
+import "./src/themes/sap_fiori_3/ToolbarSpacer.less";
+
+import "./src/themes/sap_belize/List.less";
+import "./src/themes/sap_belize_hcb/List.less";
+import "./src/themes/sap_fiori_3/List.less";
+
+import "./src/themes/sap_belize/GroupHeaderListItem.less";
+import "./src/themes/sap_belize_hcb/GroupHeaderListItem.less";
+import "./src/themes/sap_fiori_3/GroupHeaderListItem.less";
+
+import "./src/themes/sap_belize/MonthPicker.less";
+import "./src/themes/sap_belize_hcb/MonthPicker.less";
+import "./src/themes/sap_fiori_3/MonthPicker.less";
+
+import "./src/themes/sap_belize/YearPicker.less";
+import "./src/themes/sap_belize_hcb/YearPicker.less";
+import "./src/themes/sap_fiori_3/YearPicker.less";
+
+import "./src/themes/sap_belize/Popup.less";
+import "./src/themes/sap_belize_hcb/Popup.less";
+import "./src/themes/sap_fiori_3/Popup.less";
+
+import "./src/themes/sap_belize/ListItemBase.less";
+import "./src/themes/sap_belize_hcb/ListItemBase.less";
+import "./src/themes/sap_fiori_3/ListItemBase.less";
+
+import "./src/themes/sap_belize/ListItem.less";
+import "./src/themes/sap_belize_hcb/ListItem.less";
+import "./src/themes/sap_fiori_3/ListItem.less";
