+++ packages/main/package.json
@@ -44,7 +44,8 @@
     }
   },
   "scripts": {
-    "build": "npm-run-all --sequential clean lint build:templates build:samples build:bundle copy:src copy:i18n import-rewrite",
+    "build": "npm-run-all --sequential clean lint build:templates build:samples build:less copy:src build:bundle copy:i18n",
+    "build:less": "rollup -c rollup.config.less.js",
     "build:bundle": "rollup -c --environment ES5_BUILD",
     "build:templates": "mkdirp src/build/compiled && node ./lib/hbs2ui5/index.js -d src/ -o src/build/compiled",
     "build:samples": "npm run copy:test --scripts-prepend-node-path && npm run build:api --scripts-prepend-node-path && npm run build:docs --scripts-prepend-node-path && node lib/playground/index.js",
@@ -53,14 +54,13 @@
     "build:api": "jsdoc -c ../../lib/jsdoc/config.json",
     "build:docs": "node ../../lib/documentation/index.js",
     "clean": "rimraf dist src/build",
-    "copy:src": "cpx \"src/**/*.js\" dist/",
+    "copy:src": "copy-and-watch \"src/**/*.js\" \"src/**/i18n/*.json\" dist/",
     "copy:i18n": "cpx \"src/i18n/*.json\" dist/i18n",
     "copy:test": "cpx \"test/**/*.*\" dist/test-resources",
     "copy:pages": "cpx \"test/**/pages/*.*\" dist/test-resources",
     "copy:qunit": "cpx \"test/**/qunit/*.*\" dist/test-resources",
-    "dev": "npm-run-all --parallel watch:templates watch:samples watch:pages watch:qunit watch:bundle",
+    "dev": "npm-run-all --parallel watch:less watch:templates watch:samples watch:pages watch:qunit watch:src watch:bundle",
     "start": "npm-run-all --parallel serve:static dev",
-    "import-rewrite": "node ./lib/less-to-json-imports/index.js dist/",
     "lint": "eslint .",
     "test:wdio": "npm-run-all --parallel --race serve:static test:wdio-run",
     "serve:static-prepare": "cpx \"serve.json\" dist/",
@@ -69,7 +69,9 @@
     "test:wdio-run": "cross-env WDIO_LOG_LEVEL=error FORCE_COLOR=0 wdio wdio.conf.js",
     "test:jest": "jest",
     "test": "npm-run-all --sequential test:jest test:wdio",
-    "watch:bundle": "rollup -c -w --environment ES5_BUILD,DEV",
+    "watch:src": "copy-and-watch --watch \"src/**/*.js\" \"src/**/i18n/*.json\" dist/",
+    "watch:less": "rollup -c rollup.config.less.js -w",
+    "watch:bundle": "wait-on dist/less-bundle/entries-less.esm.js && rollup -c -w --environment ES5_BUILD,DEV",
     "watch:templates": "npm-watch build:templates",
     "watch:samples": "npm-watch build:samples",
     "watch:pages": "npm-watch build:pages",
@@ -97,6 +99,7 @@
     "@wdio/sync": "^5.4.13",
     "chromedriver": "^2.45.0",
     "clean-css": "^4.2.1",
+    "copy-and-watch": "^0.1.2",
     "cpx": "^1.5.0",
     "cross-env": "^5.2.0",
     "escodegen": "^1.11.0",
@@ -126,6 +129,7 @@
     "rollup-plugin-url": "^2.2.0",
     "rollup-plugin-visualizer": "^0.9.2",
     "serve": "^10.1.1",
+    "wait-on": "^3.2.0",
     "wdio-chromedriver-service": "^5.0.0"
   },
   "resolutions": {
